Support fractional bathrooms (half-baths)
- listings.bathrooms changed from INTEGER to NUMERIC(3,1)
- admin form offers half-bath options; route stores via parseFloat
- persisted in schema.sql and server.js startup migrations

Enables the DC bulk listing import (rooms with 1.5/2.5/3.5 baths).

+++ routes/admin.js
@@ -211,7 +211,7 @@ router.post('/listings', requireAdmin, upload.fields([{ name: 'image_files', max
       VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $15, $16, $17, $18, $19, $20, $21, $22, $23, $24, $25, $26, $27, $28, $29, $30, $31)
     `, [
       title, location || 'New York', neighborhood, city, state || 'NY',
-      parseInt(bedrooms) || 0, parseInt(bathrooms) || 1,
+      parseInt(bedrooms) || 0, parseFloat(bathrooms) || 1,
       sqft ? parseInt(sqft) : null, parseInt(price_monthly),
       status || 'coming_soon',
       available_from || null, available_to || null,
@@ -339,7 +339,7 @@ router.post('/listings/:id', requireAdmin, upload.fields([{ name: 'image_files',
       WHERE id = $32
     `, [
       title, location || 'New York', neighborhood, city, state || 'NY',
-      parseInt(bedrooms) || 0, parseInt(bathrooms) || 1,
+      parseInt(bedrooms) || 0, parseFloat(bathrooms) || 1,
       sqft ? parseInt(sqft) : null, parseInt(price_monthly),
       status || 'coming_soon',
       available_from || null, available_to || null,
